tools/enrich/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
pyyaml>=6.0
rich>=13.7
pytest>=8.0
